# Don't use mem::uninitialized