Don't use `mem::uninitialized()` to avoid UB and imminent deprecation. Adds a `Default` bound on the type parameter. 
The benchmark has a pretty weird regression for the `bench_vector` benchmark (it's weird because it [does not use `ElasticArray`](https://github.com/paritytech/elastic-array/blob/f6ec10c0f35d0f4403cb936a05e3b82bf9f92cf8/benches/elastic.rs#L59) at all) that I can't quite explain:

This PR:
```
test bench_elastic_array ... bench:       7,358 ns/iter (+/- 471)
test bench_vector        ... bench:       9,136 ns/iter (+/- 472)
```

Master:
```
test bench_elastic_array ... bench:       7,641 ns/iter (+/- 855)
test bench_vector        ... bench:       7,626 ns/iter (+/- 600)
```

(Many thanks to @niklasad1 for suggesting the right way here)